Added Gauge/Indicator
Added a gauge and indicator to match the position of the dial.
The gauge activates a colored light based on if the indicator is in the green, yellow, or red zones.
Gauge has functionality to be rotated in order to change the position of the green, yellow, and red zones.